:)
git-svn-id: svn+ssh://svn.in2p3.fr/class@199 0e7d625b-0364-4367-a6be-d5be4a48d228

+++ gui/trunk/CLASSWin.cxx
@@ -186,7 +186,7 @@ void MainWin::Start(vector<string> VFileName)
 	stringstream tmp;
 	tmp.str("");
 	tmp << "CLASSGui " ;
-	for(int i=0;i<VFileName.size();i++)
+	for(int i=0; i < (int)VFileName.size();i++)
 		tmp <<VFileName[i]<<" " ;
 
 	this->SetWindowName(tmp.str().c_str());
